# backend/docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "service.LoginRequest": {
        "properties": {
          "authType": {
            "description": "local, ldap",
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "输入格式: Bearer {session_id}，session_id 从登录接口获取 (Swagger 暂不支持 Cookie 自动发送，请手动输入)",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@example.com",
      "name": "API Support",
      "url": "http://www.example.com/support"
    },
    "description": "DevOps 运维平台接口文档，登录后在 Authorize 中输入 Bearer {session_id} 即可调用接口",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "DevOps 运维平台 API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/audit/cleanup": {
      "post": {
        "description": "立即执行一次过期审计日志清理",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "成功"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "清理失败"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "手动清理过期审计日志",
        "tags": [
          "审计管理"
        ]
      }
    },
    "/audit/export": {
      "get": {
        "description": "按条件导出审计日志CSV",
        "parameters": [
          {
            "description": "用户ID",
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "用户名",
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "操作类型",
            "in": "query",
            "name": "operation",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "资源路径",
            "in": "query",
            "name": "resource",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "关键词（长度\u003e=3生效，跨字段匹配）",
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "开始时间",
            "in": "query",
            "name": "startAt",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "结束时间",
            "in": "query",
            "name": "endAt",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "导出上限",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "CSV内容"
          },
          "400": {
            "content": {
              "text/csv": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "导出失败"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "导出审计日志CSV",
        "tags": [
          "审计管理"
        ]
      }
    },
    "/audit/list": {
      "get": {
        "description": "按条件分页查询审计日志",
        "parameters": [
          {
            "description": "用户ID",
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "用户名",
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "操作类型",
            "in": "query",
            "name": "operation",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "资源路径",
            "in": "query",
            "name": "resource",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "关键词（长度\u003e=3生效，跨字段匹配）",
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "开始时间",
            "in": "query",
            "name": "startAt",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "结束时间",
            "in": "query",
            "name": "endAt",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "页码",
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "每页数量",
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "查询失败"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取审计日志列表",
        "tags": [
          "审计管理"
        ]
      }
    },
    "/auth/oidc/callback": {
      "get": {
        "description": "处理 OIDC 登录回调",
        "parameters": [
          {
            "description": "Authorization Code",
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "登录成功"
          }
        },
        "summary": "OIDC 回调",
        "tags": [
          "认证管理"
        ]
      }
    },
...
    "/user/login": {
      "post": {
        "description": "用户登录（支持本地、LDAP），登录成功返回 session_id (Cookie \u0026 Body)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/service.LoginRequest"
              }
            }
          },
          "description": "登录信息",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "登录成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "认证失败"
          }
        },
        "summary": "用户登录",
        "tags": [
          "认证管理"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1"
    },
    {
      "url": "https://localhost:8000/api/v1"
    }
  ]
}